pyproject.toml
[project]
name = "fpdf-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf2>=2.8.4",
]
